retail_api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0.0
httpx>=0.27.0
sqlalchemy>=2.0.0
passlib[bcrypt]
python-jose[cryptography]
email-validator>=2.0.0
